Bring the undo manager into the container class
This allows the use of self as the target, which means the value setter will be called when undoing, which will cause redos to be registered. Registering an undo while performing an undo is registered as a redo (the undo of an undo is a redo).

Thanks to Mike Abdullah for walking me through UndoManager. 🙂